equipment resets are in, room and containers still not :)

+++ commands/command.rb
@@ -68,7 +68,7 @@ class Command
         @priority = new_attr_hash[:priority].to_i
         @keywords = new_attr_hash[:keywords].to_s.split(",")
         @keywords = [""] if @keywords.empty?
-        @lag = new_attr_hash[:lag].to_i
+        @lag = new_attr_hash[:lag].to_f
         @name = new_attr_hash[:name].to_s
         @usable_in_combat = !(new_attr_hash[:usable_in_combat].to_i.zero?)
         @creation_points = new_attr_hash[:creation_points]
